refactor: PathService 분리
