+++ src/main/java/wooteco/subway/admin/controller/PathController.java
@@ -7,15 +7,15 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import wooteco.subway.admin.dto.ShortestDistanceResponse;
-import wooteco.subway.admin.service.LineService;
+import wooteco.subway.admin.service.PathService;
 
 @RequestMapping("/paths")
 @RestController
 public class PathController {
-    private LineService lineService;
+    private PathService pathService;
 
-    public PathController(LineService lineService) {
-        this.lineService = lineService;
+    public PathController(PathService pathService) {
+        this.pathService = pathService;
     }
 
     @GetMapping
@@ -23,7 +23,7 @@ public class PathController {
         @RequestParam("source") String source,
         @RequestParam("target") String target,
         @RequestParam("type") String type) {
-        ShortestDistanceResponse shortestDistanceResponse = lineService.searchShortestDistancePath(
+        ShortestDistanceResponse shortestDistanceResponse = pathService.searchShortestDistancePath(
             source, target, type);
         return ResponseEntity.ok().body(shortestDistanceResponse);
     }
